Remove old Type component, clean up
