--- components/Type2/type2.module.scss
@@ -1,19 +1,0 @@
-.type {
-    .suffix {
-        // color: red !important;
-        color: transparent !important;
-    }
-}
-
-// Animations
-@keyframes blink {
-
-    from,
-    to {
-        color: transparent;
-    }
-
-    50% {
-        color: inherit;
-    }
-}
